requirements.txt
numpy==2.2.6
matplotlib==3.10.6
cupy-cuda13x==13.6.0